fix enemy collision

+++ src/entities/enemies/Puropen.java
@@ -8,7 +8,7 @@ import util.Utils;
 
 public class Puropen extends Enemy {
 
-	public Puropen(float posX, float posY, int width, int height, int speed, String src) {
+	public Puropen(int posX, int posY, int width, int height, int speed, String src) {
 		super(posX, posY, 47, 47, 1, null);
 		this.health = 1;
 		this.direction = "left";
@@ -49,7 +49,7 @@ public class Puropen extends Enemy {
 				break;
 		}
 		this.isAnimated = true;
-		//this.setAnimation(direction);
+		// this.setAnimation(direction);
 	}
 
 	public void render(Graphics2D g2d) {
